pyproject.toml
[tool.poetry]
name = "FastAPI Server"
version = "0.1.0"
description = ""
authors = ["Bohdan Sizomin"]

[tool.poetry.dependencies]
python = "^3.10"
setuptools = "^65.4.1"
python-multipart = "^0.0.5"
fastapi = "^0.106.0"
python-dotenv = "^1.0.0"
fastapi-pagination = "^0.12.14"
pydantic = "^2.5.3"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
sqlalchemy = "^2.0.23"
uvicorn = "^0.25.0"
alembic = "^1.13.1"
alchemical = "^1.0.1"
pyjwt = "^2.8.0"
passlib = "^1.7.4"
bcrypt = "4.0.1"
ipython = "^8.19.0"
httpx = "^0.26.0"
pydantic-settings = "^2.1.0"
email-validator = "^2.1.0.post1"
aiosqlite = "^0.19.0"
jinja2 = "^3.1.2"

[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"
black = "^22.10.0"
pytest = "^7.1.3"
faker = "^21.0.0"
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
